ROOTFS_RO: Have stage2 consider read-only or mutable rootfs (see README.md)
Additionally, we have CUSTOM_IOT as well, which might be a little dirty ..

+++ stage2/01-sys-tweaks/01-run.sh
@@ -1,5 +1,8 @@
 #!/bin/bash -e
 
+if [ "${ROOTFS_RO}" = "1" ] ; then
+    install -v -m 644 files/fstab-rootfs_ro "${ROOTFS_DIR}/etc/fstab"
+fi
 install -m 644 files/overlay_mount.service	"${ROOTFS_DIR}/lib/systemd/system/"
 install -m 755 files/overlay_mount	        "${ROOTFS_DIR}/etc/init.d/"
 
@@ -39,27 +42,22 @@ on_chroot << EOF
     else
         systemctl disable ssh
     fi
-    systemctl enable regenerate_ssh_host_keys
+    if [ "${ROOTFS_RO}" = "1" ] ; then
+        systemctl disable regenerate_ssh_host_keys
+        systemctl mask regenerate_ssh_host_keys
+    else
+        systemctl enable regenerate_ssh_host_keys
+    fi
 
-    #mkdir -p /data/overlay_a
-    #cd /data/overlay_a
-    #mkdir -p etc/upper etc/work home/upper home/work srv/upper srv/work tmp/upper tmp/work var/upper var/work
-    #
-    #mkdir -p /data/overlay_b
-    #cd /data/overlay_b
-    #mkdir -p etc/upper etc/work home/upper home/work srv/upper srv/work tmp/upper tmp/work var/upper var/work
-    #
-    #cd /data
-    #ln -s overlay_a overlay
-    #
+    if [ "${ROOTFS_RO}" = "1" ] ; then
+        sed -i -e 's/^D \/tmp/#D \/tmp/g' /usr/lib/tmpfiles.d/tmp.conf
 
-    sed -i -e 's/^D \/tmp/#D \/tmp/g' /usr/lib/tmpfiles.d/tmp.conf
-
-    systemctl unmask overlay_mount
-    systemctl enable overlay_mount
-
-    #systemctl disable overlay_mount
-    #systemctl mask overlay_mount
+        systemctl unmask overlay_mount
+        systemctl enable overlay_mount
+    else
+        systemctl disable overlay_mount
+        systemctl mask overlay_mount
+    fi
 
     # Setup unique system folder /boot/sys_arm64_000
     mkdir -p /boot/sys_arm64_000
@@ -85,46 +83,60 @@ on_chroot << EOF
         fi
     done
 
-    systemctl disable resize2fs_once
-    systemctl mask resize2fs_once
+    if [ "${ROOTFS_RO}" = "1" ] ; then
+        systemctl disable resize2fs_once
+        systemctl mask resize2fs_once
 
-    systemctl disable apt-daily
-    systemctl mask apt-daily
-    systemctl disable apt-daily.timer
-    systemctl mask apt-daily.timer
-    systemctl disable apt-daily-upgrade.timer
-    systemctl mask apt-daily-upgrade.timer
+        systemctl disable apt-daily
+        systemctl mask apt-daily
+        systemctl disable apt-daily.timer
+        systemctl mask apt-daily.timer
+        systemctl disable apt-daily-upgrade.timer
+        systemctl mask apt-daily-upgrade.timer
 
-    systemctl disable logrotate.timer
-    systemctl mask logrotate.time
-    systemctl disable man-db.timer
-    systemctl mask man-db.timer
+        systemctl disable logrotate.timer
+        systemctl mask logrotate.time
+        systemctl disable man-db.timer
+        systemctl mask man-db.timer
 
-    systemctl disable bluetooth
-    systemctl mask bluetooth
+        sed -i -e 's/#Storage=auto/Storage=volatile/g;s/#Compress=yes/Compress=yes/g;s/#RuntimeMaxUse=/RuntimeMaxUse=1M/g;s/#ForwardToSyslog=yes/ForwardToSyslog=no/g;s/#ForwardToWall=yes/ForwardToWall=no/g' /etc/systemd/journald.conf
 
-    sed -i -e 's/#Storage=auto/Storage=volatile/g;s/#Compress=yes/Compress=yes/g;s/#RuntimeMaxUse=/RuntimeMaxUse=1M/g;s/#ForwardToSyslog=yes/ForwardToSyslog=no/g;s/#ForwardToWall=yes/ForwardToWall=no/g' /etc/systemd/journald.conf
+            sed -i -e 's/MODULES=most/MODULES=dep/g;s/BUSYBOX=auto/BUSYBOX=y/g' /etc/initramfs-tools/initramfs.conf
 
-    sed -i -e 's/MODULES=most/MODULES=dep/g;s/BUSYBOX=auto/BUSYBOX=y/g' /etc/initramfs-tools/initramfs.conf
+        echo "squashfs"     >> /etc/modules
+        echo "squashfs"     >> /etc/initramfs-tools/modules
+    else
+        systemctl enable resize2fs_once
+    fi
 
-    echo "squashfs"     >> /etc/modules
-    echo "i2c-bcm2708"  >> /etc/modules
-    # echo "i2c-dev"    >> /etc/modules
-    # echo "rtc-ds1307" >> /etc/modules
+    if [ "${CUSTOM_IOT}" = "1" ] ; then
+        systemctl disable bluetooth
+        systemctl mask bluetooth
 
-    echo "squashfs"     >> /etc/initramfs-tools/modules
-    echo "i2c-bcm2708"  >> /etc/initramfs-tools/modules
-    # echo "i2c-dev"    >> /etc/initramfs-tools/modules
-    # echo "rtc-ds1307" >> /etc/initramfs-tools/modules
+        echo "i2c-bcm2708"  >> /etc/modules
+        # echo "i2c-dev"    >> /etc/modules
+        # echo "rtc-ds1307" >> /etc/modules
+
+        if [ "${ROOTFS_RO}" = "1" ] ; then
+            echo "i2c-bcm2708"  >> /etc/initramfs-tools/modules
+            # echo "i2c-dev"    >> /etc/initramfs-tools/modules
+            # echo "rtc-ds1307" >> /etc/initramfs-tools/modules
+        fi
+    fi
 EOF
 
-install -m 644 files/boot/sys_arm64_000/cmdline.txt  "${ROOTFS_DIR}/boot/sys_arm64_000/"
+if [ "${ROOTFS_RO}" = "1" ] ; then
+    install -m 644 files/boot/config-rootfs_ro.txt 	     "${ROOTFS_DIR}/boot/"
+    install -m 644 files/boot/config-rootfs_ro.txt 	     "${ROOTFS_DIR}/boot/sys_arm64_000/"
+    install -m 644 files/boot/sys_arm64_000/cmdline-rootfs_ro.txt  "${ROOTFS_DIR}/boot/sys_arm64_000/cmdline.txt"
 
-install -m 644 files/boot/config.txt 	             "${ROOTFS_DIR}/boot/"
-install -m 644 files/boot/config.txt 	             "${ROOTFS_DIR}/boot/sys_arm64_000/"
-
-install -m 755 files/initramfs/loop_rootfs 	"${ROOTFS_DIR}/etc/initramfs-tools/scripts/init-premount/"
-install -m 755 files/initramfs/fsck_custom 	"${ROOTFS_DIR}/etc/initramfs-tools/hooks/"
+    install -m 755 files/initramfs/loop_rootfs 	"${ROOTFS_DIR}/etc/initramfs-tools/scripts/init-premount/"
+    install -m 755 files/initramfs/fsck_custom 	"${ROOTFS_DIR}/etc/initramfs-tools/hooks/"
+else
+    install -m 644 files/boot/config-rootfs_rw.txt 	     "${ROOTFS_DIR}/boot/"
+    install -m 644 files/boot/config-rootfs_rw.txt 	     "${ROOTFS_DIR}/boot/sys_arm64_000/"
+    install -m 644 files/boot/sys_arm64_000/cmdline-rootfs_rw.txt  "${ROOTFS_DIR}/boot/sys_arm64_000/cmdline.txt"
+fi
 
 if [ "${USE_QEMU}" = "1" ]; then
 	echo "enter QEMU mode"
@@ -143,10 +155,14 @@ on_chroot <<EOF
     setupcon --force --save-only -v
     usermod --pass='*' root
 
-    KVERSION=\$(ls /lib/modules/ | tail -n 1)
-    echo "mkinitramfs for kernel version: \${KVERSION}"
-    /usr/sbin/mkinitramfs -o /boot/sys_arm64_000/initrd.img \${KVERSION}
+    if [ "${ROOTFS_RO}" = "1" ] ; then
+        KVERSION=\$(ls /lib/modules/ | tail -n 1)
+        echo "mkinitramfs for kernel version: \${KVERSION}"
+        /usr/sbin/mkinitramfs -o /boot/sys_arm64_000/initrd.img \${KVERSION}
+    fi
 EOF
 
-rm -f "${ROOTFS_DIR}/etc/ssh/"ssh_host_*_key*
+if [ "${ROOTFS_RO}" != "1" ] ; then
+    rm -f "${ROOTFS_DIR}/etc/ssh/"ssh_host_*_key*
+fi
